Reacted on -Wsign-compare warnings of gcc
git-svn-id: http://svn.libburnia-project.org/libburn/trunk@3792 fff8544e-141b-0410-a922-c1ea1f00b384

+++ libburn/structure.c
@@ -33,7 +33,7 @@ extern struct libdax_msgs *libdax_messenger;
 		return 0;\
 	if (pos == BURN_POS_END)\
 		pos = TO->NEW##s;\
-	if (pos > TO->NEW##s)\
+	if ((int) pos > TO->NEW##s)\
 		return 0;\
 \
 	tmp = realloc(TO->NEW, sizeof(struct NEW *) * (TO->NEW##s + 1));\
